+++ CHANGELOG.md
@@ -9,3 +9,4 @@ have been made so far, between releases.
       gained no boost from __slots__ being specified for them whatsoever
 * fixed __str__ in `AMQPContentPropertyList`
 * added __slots__ to `AMQPFrame`
+* removed redundant logging in `coolamqp.uplink.connection.connection`